Code clean up and moving, used plugins plugin, updated gemspec and changelog
