--- doc/untitled.html
@@ -1,138 +1,0 @@
-<style type="text/css" media="screen">
-  table {
-    empty-cells: show;
-    border-collapse: collapse;
-  }
-  td {
-    min-width: 50px;
-    text-align: center;
-    font-family: Arial;
-    background-color: #eee;
-    border: dashed 1px #000;
-  }
-</style>
-
-<table cellspacing="0" cellpadding="5">
-  <tr>
-    <td></td>
-    <td colspan='3'>a</td>
-    <td colspan='3'>b</td>
-    <td colspan='3'>c</td>
-    <td colspan='3'></td>
-  </tr>
-  <tr>
-    <td rowspan='3'>d</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>s</td>
-    <td rowspan='3'>s</td>
-    <td rowspan='6'>s</td>
-  </tr>
-  <tr>
-    <!-- <td rowspan='3'>d</td> -->
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>s</td>
-    <!-- <td rowspan='3'>s</td> -->
-    <!-- <td rowspan='6'>s</td> -->
-  </tr>
-  <tr>
-    <!-- <td rowspan='3'>d</td> -->
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>s</td>
-    <!-- <td rowspan='3'>s</td> -->
-    <!-- <td rowspan='6'>s</td> -->
-  </tr>
-  <tr>
-    <td rowspan='3'>e</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>s</td>
-    <td rowspan='3'>s</td>
-    <!-- <td rowspan='6'>s</td> -->
-  </tr>
-  <tr>
-    <!-- <td rowspan='3'>e</td> -->
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>s</td>
-    <!-- <td rowspan='3'>s</td> -->
-    <!-- <td rowspan='6'>s</td> -->
-  </tr>
-  <tr>
-    <!-- <td rowspan='3'>e</td> -->
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>x</td>
-    <td>y</td>
-    <td>z</td>
-    <td>s</td>
-    <!-- <td rowspan='3'>s</td> -->
-    <!-- <td rowspan='6'>s</td> -->
-  </tr>
-  <tr>
-    <td></td>
-    <td>s</td>
-    <td>s</td>
-    <td>s</td>
-    <td>s</td>
-    <td>s</td>
-    <td>s</td>
-    <td>s</td>
-    <td>s</td>
-    <td>s</td>
-    <td colspan='3'></td>
-  </tr>
-  <tr>
-    <td></td>
-    <td colspan='3'>s</td>
-    <td colspan='3'>s</td>
-    <td colspan='3'>s</td>
-    <td colspan='3'></td>
-  </tr>
-  <tr>
-    <td></td>
-    <td colspan='9'>s</td>
-    <td colspan='3'></td>
-  </tr>
-</table>
