+++ lib/open_project/reporting/version.rb
@@ -1,5 +1,5 @@
 module OpenProject
   module Reporting
-    VERSION = '4.0.0'
+    VERSION = '4.0.1.pre1'
   end
 end
